refactor: 상수의 책임을 클래스에게 분할
lottoNumber.js 내부에 모든 상수를 분리함
다른 위치에서 사용되지 않는 상수는 사용하는 클래스로 책임을 분할함
