+++ __tests__/LottoTest.js
@@ -33,7 +33,7 @@ describe('로또 클래스 테스트', () => {
     );
   });
 
-  test('로또 번호에 1미만 또는 45초과의 숫자가 있으면 예외가 발생한다.', () => {
+  test('로또 번호에 실수가 있으면 예외가 발생한다.', () => {
     expect(() => new Lotto([1, 2, 3, 4, 5, 6.6])).toThrow(
       ERROR_MESSAGE.CONTAIN_FLOAT,
     );
