+++ src/constants/lottoStandard.js
@@ -1,0 +1,3 @@
+const MONEY_UNIT = 1000;
+
+export default MONEY_UNIT;
