--- src/constants/magicNumber.js
@@ -1,29 +1,0 @@
-const LOTTO_BOUNDARY = Object.freeze({ MIN: 1, MAX: 45 });
-
-const MONEY_UNIT = 1000;
-
-const WINNING_COUNT = Object.freeze({
-  FIRST_RANK: 6,
-  SECOND_THIRD_RANK: 5,
-  FOURTH_RANK: 4,
-  FIFTH_RANK: 3,
-});
-
-const RANK = Object.freeze({
-  FIRST_RANK: 1,
-  SECOND_RANK: 2,
-  THIRD_RANK: 3,
-  FOURTH_RANK: 4,
-  FIFTH_RANK: 5,
-  SIXTH_RANK: 6,
-});
-
-const RANK_MONEY = Object.freeze({
-  FIFTH_RANK: 5000,
-  FOURTH_RANK: 50000,
-  THIRD_RANK: 1500000,
-  SECOND_RANK: 30000000,
-  FIRST_RANK: 2000000000,
-});
-
-export { LOTTO_BOUNDARY, MONEY_UNIT, WINNING_COUNT, RANK, RANK_MONEY };
